refator(jest): use default directory for jest caches
The current cacheDirectory is set to `tmp` within master’s jest config.
 This creates the directory `tmp` in the project directory which
typically needs to be ignored in a .gitignore file.  The default
directory is /tmp I think.  I think the only reason I was using the tmp
directory initially was when I was testing out jest, but now it’s not
needed and causes a headache each time a new project is started.